refactor(course): ♻️ 提取共用课程模型
