+++ src/themeBasedComponents/ios/courseItem.tsx
@@ -6,7 +6,7 @@ import {
   COURSE_ITEM_HEIGHT,
   ITEM_COLORS,
 } from '@/constants/SCHEDULE';
-import { CourseTransferType } from '@/modules/courseTable/components/courseTable/type';
+import type { CourseTransferType } from '@/modules/courseTable/types';
 import useVisualScheme from '@/store/visualScheme';
 
 const CourseItem: React.FC<CourseTransferType> = props => {
