+++ src/utils/updateWidget.ts
@@ -1,7 +1,7 @@
 import { ExtensionStorage } from '@bacons/apple-targets';
 import { Platform } from 'react-native';
 
-import type { courseType } from '@/modules/courseTable/components/courseTable/type';
+import type { courseType } from '@/modules/courseTable/types';
 import useTimeStore from '@/store/time';
 import {
   buildAndroidWidgetCourseData,
